added an explicit class for the graph layout settings to unclutter the component class
